backend/requirements.txt
# ==============================================
# CORE DEPENDENCIES (Install First)
# ==============================================

# Core FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
motor==3.3.2
redis==5.0.1
pymongo==4.6.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
cryptography==42.0.0

# OAuth
authlib==1.3.0
httpx==0.26.0

# Email
aiosmtplib==5.0.0
email-validator==2.1.0
jinja2==3.1.3

# SMS & Validation
twilio==8.12.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Data Processing & Export
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.7
apscheduler==3.10.4

# AWS S3
boto3==1.34.20

# Web Scraping (for SMI)
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Push Notifications
pywebpush==1.14.0

# ==============================================
# ML DEPENDENCIES (Heavy - Install Last)
# ==============================================

# Base ML dependencies
numpy<2.0
scikit-learn==1.3.2
joblib==1.3.2

# Image Processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# PyTorch (CPU only) - Railway compatible
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu
torchaudio==2.0.0+cpu

# Whisper (after PyTorch)
openai-whisper==20231117

# FAISS & Sentence Transformers (after PyTorch)
faiss-cpu==1.7.4
sentence-transformers==2.7.0
